Move common listening_activity stuff to a separate file for reuse
